Move getSubtitle to book and make it a text comp
